[PyITA] Move GELU functions
