+++ PyITA/test_gelu.py
@@ -1,3 +1,28 @@
+# ----------------------------------------------------------------------
+#
+# File: test_gelu.py
+#
+#
+# Copyright (C) 2024, ETH Zurich and University of Bologna.
+#
+#
+# ----------------------------------------------------------------------
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the License); you may
+# not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an AS IS BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# This test file is used to check the integer quantization of the GELU function.
+
 import pytest
 import torch
 import numpy as np
@@ -6,6 +31,8 @@ import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
 
+from .util import *
+from .gelu import *
 from .ITA import *
 
 N_SAMPLES = 75
@@ -41,6 +68,8 @@ def plot(data: pd.DataFrame, title: str, quantized_y_label: str, expected_y_labe
     ax.set_xlabel('$x$')
     ax.set_ylabel('Value')
     filename = os.path.join(plot_dir, f'{title}.png')
+    if not os.path.exists(plot_dir):
+        os.makedirs(plot_dir)
     plt.savefig(filename)
